# openapi.yaml
openapi: 3.1.0
info:
  title: IA Prononciation API
  version: 1.0.0

servers:
  - url: https://ia-prononciation-server.onrender.com
    description: Serveur de production hébergé sur Render

paths:
  /analyse-prononciation:
    post:
      summary: Analyse la prononciation d'un fichier audio
      operationId: analysePrononciation
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                fichier:
                  type: string
                  format: binary
                  description: Fichier audio à analyser (.wav, .mp3)
                base64_audio:
                  type: string
                  description: Audio encodé en base64 (en alternative au fichier)
                texte_cible:
                  type: string
                  description: Phrase attendue
      responses:
        '200':
          description: Résultat de l'analyse de prononciation
          content:
            application/json:
              schema:
                type: object
                properties:
                  moteur_utilise:
                    type: string
                  langue_detectee:
                    type: string
                  transcription:
                    type: string
                  attendu:
                    type: string
                  score:
                    type: number
                  erreurs:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                  audio_url:
                    type: string
                  conseils:
                    type: array
                    items:
                      type: string

  /score:
    post:
      summary: Obtenir un score sans conseils ou audio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                fichier:
                  type: string
                  format: binary
                texte_cible:
                  type: string
      responses:
        '200':
          description: Score de la prononciation
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: number
                  message:
                    type: string
                  erreurs:
                    type: array
                    items:
                      type: string

  /exercice:
    get:
      summary: Obtenir une phrase aléatoire à prononcer
      parameters:
        - name: langue
          in: query
          required: true
          schema:
            type: string
        - name: niveau
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Phrase et audio à prononcer
          content:
            application/json:
              schema:
                type: object
                properties:
                  phrase:
                    type: string
                  audio_url:
                    type: string

  /ajouter-phrase:
    post:
      summary: Ajouter une phrase personnalisée à la base d'exercices
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                langue:
                  type: string
                niveau:
                  type: string
                phrase:
                  type: string
      responses:
        '200':
          description: Confirmation d'ajout
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /health:
    get:
      summary: Vérifie si le serveur est opérationnel
      responses:
        '200':
          description: Serveur fonctionnel
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string